medimate-ai-service/requirements.txt
# Core ML/AI Libraries
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Additional dependencies for PDF processing
pdf2image>=3.1.0

# Medical NLP Libraries
spacy>=3.6.0
scispacy>=0.5.3
medspacy>=1.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Database and Caching
redis>=4.6.0
pymongo>=4.4.0
sqlalchemy>=2.0.0

# ML Operations and Monitoring
mlflow>=2.5.0
wandb>=0.15.0
prometheus-client>=0.17.0

# Security and Encryption
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
celery>=5.3.0